+++ app/src/main/java/com/king/zxing/app/CustomCaptureActivity.java
@@ -77,7 +77,7 @@ public class CustomCaptureActivity extends CaptureActivity {
         getCameraScan().setPlayBeep(true)//设置是否播放音效，默认为false
                 .setVibrate(true)//设置是否震动，默认为false
 //                .setCameraConfig(new CameraConfig())//设置相机配置信息，CameraConfig可覆写options方法自定义配置
-                .setCameraConfig(new ResolutionCameraConfig(this))//设置CameraConfig，可以根据自己的需求去自定义配置
+//                .setCameraConfig(new ResolutionCameraConfig(this))//设置CameraConfig，可以根据自己的需求去自定义配置
                 .setNeedAutoZoom(false)//二维码太小时可自动缩放，默认为false
                 .setNeedTouchZoom(true)//支持多指触摸捏合缩放，默认为true
                 .setDarkLightLux(45f)//设置光线足够暗的阈值（单位：lux），需要通过{@link #bindFlashlightView(View)}绑定手电筒才有效
@@ -117,12 +117,10 @@ public class CustomCaptureActivity extends CaptureActivity {
     }
 
     private void showToast(String text){
-        if(toast == null){
-            toast = Toast.makeText(this,text,Toast.LENGTH_SHORT);
-        }else{
-            toast.setText(text);
-            toast.setDuration(Toast.LENGTH_SHORT);
+        if(toast != null){
+            toast.cancel();
         }
+        toast = Toast.makeText(this,text,Toast.LENGTH_SHORT);
         toast.show();
     }