+++ zxing-lite/src/main/java/com/king/zxing/analyze/ImageAnalyzer.java
@@ -5,9 +5,8 @@ import android.content.res.Configuration;
 import android.graphics.ImageFormat;
 
 import com.google.zxing.Result;
+import com.king.zxing.util.BitmapUtils;
 import com.king.zxing.util.LogUtils;
-
-import java.nio.ByteBuffer;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -15,12 +14,14 @@ import androidx.camera.core.ImageProxy;
 
 /**
  * 图像分析器
+ *
  * @author <a href="mailto:jenly1314@gmail.com">Jenly</a>
  */
 public abstract class ImageAnalyzer implements Analyzer {
 
     /**
      * 分析图像数据
+     *
      * @param data
      * @param width
      * @param height
@@ -29,25 +30,23 @@ public abstract class ImageAnalyzer implements Analyzer {
     public abstract Result analyze(byte[] data, int width, int height);
 
     @Override
-    public Result analyze(@NonNull ImageProxy image,int orientation) {
-        if(image.getFormat() == ImageFormat.YUV_420_888){
-            @SuppressLint("UnsafeExperimentalUsageError")
-            ByteBuffer buffer = image.getPlanes()[0].getBuffer();
-            byte[] data = new byte[buffer.remaining()];
-            buffer.get(data);
+    public Result analyze(@NonNull ImageProxy image, int orientation) {
+        if (image.getFormat() == ImageFormat.YUV_420_888) {
             int width = image.getWidth();
             int height = image.getHeight();
-            if(orientation == Configuration.ORIENTATION_PORTRAIT){
+            @SuppressLint("UnsafeOptInUsageError")
+            byte[] data = BitmapUtils.yuv420ThreePlanesToNV21(image.getImage().getPlanes(), width, height).array();
+            if (orientation == Configuration.ORIENTATION_PORTRAIT) {
                 byte[] rotatedData = new byte[data.length];
                 for (int y = 0; y < height; y++) {
-                    for (int x = 0; x < width; x++){
+                    for (int x = 0; x < width; x++) {
                         rotatedData[x * height + height - y - 1] = data[x + y * width];
                     }
                 }
-                return analyze(rotatedData,height,width);
+                return analyze(rotatedData, height, width);
             }
-            return analyze(data,width,height);
-        }else{
+            return analyze(data, width, height);
+        } else {
             LogUtils.w("imageFormat: " + image.getFormat());
         }
         return null;