+++ zxing-lite/src/main/java/com/king/zxing/config/AspectRatioCameraConfig.java
@@ -3,7 +3,10 @@ package com.king.zxing.config;
 import android.content.Context;
 import android.util.DisplayMetrics;
 
+import com.king.zxing.CameraScan;
 import com.king.zxing.util.LogUtils;
+
+import java.util.Locale;
 
 import androidx.annotation.NonNull;
 import androidx.camera.core.AspectRatio;
@@ -18,6 +21,9 @@ import androidx.camera.core.Preview;
  */
 public final class AspectRatioCameraConfig extends CameraConfig {
 
+    /**
+     * 纵横比
+     */
     private int mAspectRatio;
 
     public AspectRatioCameraConfig(Context context) {
@@ -34,14 +40,15 @@ public final class AspectRatioCameraConfig extends CameraConfig {
         DisplayMetrics displayMetrics = context.getResources().getDisplayMetrics();
         int width = displayMetrics.widthPixels;
         int height = displayMetrics.heightPixels;
+        LogUtils.d(String.format(Locale.getDefault(), "displayMetrics: %dx%d", width, height));
 
-        float ratio = Math.max(width, height) / Math.min(width, height);
-        if (Math.abs(ratio - 4.0F / 3.0F) < Math.abs(ratio - 16.0F / 9.0F)) {
+        float ratio = Math.max(width, height) / (float) Math.min(width, height);
+        if (Math.abs(ratio - CameraScan.ASPECT_RATIO_4_3) < Math.abs(ratio - CameraScan.ASPECT_RATIO_16_9)) {
             mAspectRatio = AspectRatio.RATIO_4_3;
         } else {
             mAspectRatio = AspectRatio.RATIO_16_9;
         }
-        LogUtils.d("aspectRatio:" + mAspectRatio);
+        LogUtils.d("aspectRatio: " + mAspectRatio);
     }
 
     @NonNull